:recycle: refactor management commands, with base class
