+++ quant_candles/serializers/base.py
@@ -28,7 +28,7 @@ class BaseParameterSerializer(serializers.Serializer):
         time_ago = data.pop("time_ago")
         now = get_current_time()
         timestamp_to = get_min_time(now, "1t")
-        timestamp_from = timestamp_to - pd.Timedelta(time_ago)
+        timestamp_from = timestamp_to - time_ago
         data["timestamp_to"] = timestamp_to
         data["timestamp_from"] = timestamp_from
         return data
